test/basic: use wrap-file for c-meson-cargo
Avoid the direct subproject symlink and use a wrap-file with a relative
path instead. This allows running `meson dist` since the test is now
fully self-contained when vendored.

This requires changes to be committed to be seen by the test. But this
seems like a low price to pay to get dist-support in the tests.

Signed-off-by: David Rheinsberg <david@readahead.eu>